UBER-769 Extract tiptap extensions into separate package
