+++ dev/prod/package.json
@@ -47,6 +47,7 @@
     "@hcengineering/platform": "^0.6.9",
     "@hcengineering/ui": "^0.6.10",
     "@hcengineering/theme": "^0.6.3",
+    "@hcengineering/tiptap": "^0.6.0",
     "svelte": "3.55.1",
     "@hcengineering/login": "^0.6.7",
     "@hcengineering/login-assets": "^0.6.0",
